1) php artisan make:controller CategoryController --resource  - categories.search  - categories.index  - categories.show 2) main SEARCH представление 3) MasterController - masters.index - masters.show - masters.create - (store) -> master.management - master.meetings - (update) -> master.manegement - (updateMaster) ->master.management - (destroyMaster) -> master.management 4)  MeetingController проверка нужна на мастера (только Client может записаться) 5) User - is_admin поле для проверки и регистрации юзеров, добавили в МИГРАЦИИ 6) MasterFactory - обработка фото, загрузка и добавление, добавили в МИГРАЦИИ 7) ServiceFactory - переделали прайс на INTEGER 8) сделали представления

+++ database/migrations/2025_02_15_083625_create_services_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('title')->nullable(false);
-            $table->float('price')->nullable(false);
+            $table->integer('price')->nullable(false);
+            //$table->string('is_admin')->always('user');
             $table->foreignId('category_id')->constrained();
             $table->foreignId('master_id')->constrained();
-
             // CONSTRAINT - ключевое слово, которое указывает, что в данной секции описывается ограничение, 
             // которое налагается на данные в таблице, и которое будет проверяться подсистемой контроля целостности 
             // и непротиворечивости данных сервера, что в свою очередь не позволит внести или изменить данные в таблице так,
